// src/api/api-course/models/Finances.ts
export interface Finances {
    /**
     * 
     * @type {FinancesId}
     * @memberof Finances
     */
    id: FinancesId;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    investorGave?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    investorReturn?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    spentBudget?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    toysPlanned?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    toysSold?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    toysLeft?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    membersBudget?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    profit?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    expenseAmount?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    calculatedBudget?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    itemsCost?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    recommendedPrice?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    pricePerItem?: number;
    /**
     * 
     * @type {number}
     * @memberof Finances
     */
    soldPrice?: number;
}